Remove tmp query + resolver and move colated module tests to top level
