+++ src/modules/logout/logout.test.ts
@@ -1,9 +1,9 @@
 import * as faker from 'faker';
 import { Connection } from 'typeorm';
 
-import { User } from '../../../entity/User';
-import { createTestConn } from '../../../test/createTestConn';
-import { TestClient } from '../../../test/TestClient';
+import { User } from '../../entity/User';
+import { createTestConn } from '../../test/createTestConn';
+import { TestClient } from '../../test/TestClient';
 
 faker.seed(Date.now() + Math.random());
